% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figlet.R
\name{figlet}
\alias{figlet}
\title{figlet}
\usage{
figlet(msg)
}
\arguments{
\item{msg}{message}
}
\value{
characters
}
\description{
figlet
}
\details{
should have figlet installed in your system
}
\author{
guangchuang yu
}
