% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lolcat.R
\name{lolcat}
\alias{lolcat}
\title{lolcat}
\usage{
lolcat(msg)
}
\arguments{
\item{msg}{message}
}
\value{
characters
}
\description{
lolcat
}
\details{
should have lolcat installed (pip install lolcat)
}
\author{
guangchuang yu
}
