% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figlet.R, R/lolcat.R
\docType{methods}
\name{show}
\alias{figlet-class}
\alias{lolcat-class}
\alias{show}
\alias{show,figlet-method}
\alias{show,lolcat-method}
\title{show method}
\usage{
\S4method{show}{figlet}(object)

\S4method{show}{lolcat}(object)
}
\arguments{
\item{object}{figlet object}
}
\value{
message
}
\description{
show method for figlet instance

show method for lolcat instance
}

