% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ydict.R
\name{ydict}
\alias{ydict}
\title{ydict}
\usage{
ydict(word)
}
\arguments{
\item{word}{input word}
}
\value{
formatted outpuut queried from youdao
}
\description{
translate English to Chinese and Chinese to English, using youdao dictionary
}
\references{
\url{https://github.com/TimothyYe/ydict}
}
\author{
Guangchuang Yu
}
